--- app/src/main/java/com/example/beattrack/SplashScreen.java
@@ -1,24 +1,0 @@
-package com.example.beattrack;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-public class SplashScreen extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash_screen);
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent= new Intent(SplashScreen.this,MainActivity.class);
-                startActivity(intent);
-                finish();
-            }
-        },3500);
-    }
-}
