+++ app/src/main/java/com/example/beattrack/start.java
@@ -1,18 +1,19 @@
-package com.example.beattrack_4750;
-
-import androidx.appcompat.app.AppCompatActivity;
+package com.example.beattrack;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 
-public class start extends AppCompatActivity {
+import androidx.appcompat.app.AppCompatActivity;
 
+
+public class start extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_start);
     }
+
     public void onSignupClicked(View view) {
         Intent intent = new Intent(this, signup.class);
         startActivity(intent);
