+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">BeatTrack_4750</string>
+    <string name="app_name">BeatTrack</string>
     <string name="track_your_heartbeat_embrace_your_rhythm">Track Your Heartbeat, Embrace Your Rhythm</string>
     <string name="name">Name</string>
     <string name="height">Height</string>
