+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "BeatTrack_4750"
+rootProject.name = "BeatTrack"
 include ':app'
